fix: useSearch 분리 및 모드 설정 적용
